+++ packages/terra-status-view/tests/jest/StatusView.test.jsx
@@ -1,14 +1,15 @@
 import React from 'react';
 import Button from 'terra-button';
+/* eslint-disable-next-line import/no-extraneous-dependencies */
+import { mountWithIntl } from 'terra-enzyme-intl';
 import IconDiamond from 'terra-icon/lib/icon/IconDiamond';
-import intlContexts from './intl-context-setup';
 import StatusView, { StatusViewVariants } from '../../src/StatusView';
 
 // Snapshot Tests
 it('should render an image with a default status view', () => {
   const statusView = (<StatusView />);
 
-  const wrapper = mount(statusView, intlContexts.mountContext);
+  const wrapper = mountWithIntl(statusView);
   expect(wrapper).toMatchSnapshot();
 });
 
@@ -19,7 +20,7 @@ it('should render a single child element', () => {
     </StatusView>
   );
 
-  const wrapper = mount(statusView, intlContexts.mountContext);
+  const wrapper = mountWithIntl(statusView);
   expect(wrapper).toMatchSnapshot();
 });
 
@@ -31,41 +32,41 @@ it('should render an image with a no-data status view with all message props giv
     </StatusView>
   );
 
-  const wrapper = mount(statusView, intlContexts.mountContext);
+  const wrapper = mountWithIntl(statusView);
   expect(wrapper).toMatchSnapshot();
 });
 
 it('should render an image with a no-data status view that will hide the glyph from being shown', () => {
   const statusView = (<StatusView variant={StatusViewVariants.NODATA} isGlyphHidden />);
 
-  const wrapper = mount(statusView, intlContexts.mountContext);
+  const wrapper = mountWithIntl(statusView);
   expect(wrapper).toMatchSnapshot();
 });
 
 it('should render an image with a no-matching-results status-view that will force the content to align at the top', () => {
   const statusView = (<StatusView variant={StatusViewVariants.NOMATCHINGRESULTS} isAlignedTop />);
 
-  const wrapper = mount(statusView, intlContexts.mountContext);
+  const wrapper = mountWithIntl(statusView);
   expect(wrapper).toMatchSnapshot();
 });
 
 it('should render an image with a not-authorized status-view', () => {
   const statusView = (<StatusView variant={StatusViewVariants.NOTAUTHORIZED} />);
 
-  const wrapper = mount(statusView, intlContexts.mountContext);
+  const wrapper = mountWithIntl(statusView);
   expect(wrapper).toMatchSnapshot();
 });
 
 it('should render an image with an error status-view', () => {
   const statusView = (<StatusView variant={StatusViewVariants.ERROR} />);
 
-  const wrapper = mount(statusView, intlContexts.mountContext);
+  const wrapper = mountWithIntl(statusView);
   expect(wrapper).toMatchSnapshot();
 });
 
 it('should render an image with a custom status-view', () => {
   const statusView = (<StatusView variant={StatusViewVariants.CUSTOM} customGlyph={<IconDiamond />} />);
 
-  const wrapper = mount(statusView, intlContexts.mountContext);
+  const wrapper = mountWithIntl(statusView);
   expect(wrapper).toMatchSnapshot();
 });