+++ packages/terra-overlay/CHANGELOG.md
@@ -3,12 +3,16 @@ Changelog
 
 Unreleased
 ----------
+
+3.7.0 - (March 21, 2019)
+------------------
 ### Added
 * Added overlay prop to OverlayContainer to separate overlay from container-content
 * Added condition in overlay setNode method to prevent screen reader from reading container-content
 
 ### Changed
 * Replaced local keyCode values with keycode-js npm package
+* Update Jest tests to use terra-enzyme-intl
 
 3.6.0 - (February 26, 2019)
 ------------------